fix(numerics): int multiplications overflowed before being promoted to double
CodeQL reports 92 cs/loss-of-precision errors, but they are not this branch's:
the alerts are recorded against refs/heads/master, and the PR touches only 4 of
the 54 files they live in. The other 50 files are untouched here, so fixing them
inside a model-shard PR would be unrelated churn in code this branch never
modified. They are master's debt and worth a pass of their own.

Fixed the ones this PR is actually responsible for:

  GOLEMAlgorithm        sigmaSquared = rss / (n * d)
                        score = (n * d / 2.0) * Math.Log(...)
  NelsonAalenEstimator  (double)eventsAtTime / (atRisk * atRisk)

In each, the multiplication is evaluated in int and can wrap BEFORE the division
promotes it, so a large graph or cohort divides by a wrapped -- possibly negative
-- count rather than the intended magnitude. Casting one operand first makes the
whole expression double.

OnlineLearningModelBase is a different case and is fixed differently:

  Math.Pow(0.5, SampleCount / 10000)

The truncation there is DELIBERATE -- it is what makes StepDecay a step decay,
halving once per 10,000 samples and staying flat between steps. Writing 10000.0
would have silently converted it to a smooth exponential and changed every
schedule selecting StepDecay. The integer division is kept and made explicit with
a cast, so it reads as intent rather than as an accidental int/int.

The fifth alert, in TabDPTRegression, no longer exists: the line it names was
already changed on this branch.

Build clean; 172 tests across the touched areas pass (GOLEMAlgorithmTests 14/14
in isolation -- a "1 ms" failure in the batched run was scheduler contention, not
these changes).

+++ src/CausalDiscovery/ContinuousOptimization/GOLEMAlgorithm.cs
@@ -200,7 +200,10 @@ public class GOLEMAlgorithm<T> : ContinuousOptimizationBase<T>
             }
         }
 
-        double sigmaSquared = rss / (n * d);
+        // (double) on one operand FIRST: n * d is an int multiply that overflows
+        // before the division promotes it, so a large graph silently divides by a
+        // negative or wrapped count.
+        double sigmaSquared = rss / ((double)n * d);
         if (sigmaSquared < 1e-15) sigmaSquared = 1e-15;
 
         // I - W
@@ -215,7 +218,8 @@ public class GOLEMAlgorithm<T> : ContinuousOptimizationBase<T>
         double logDetImW = ComputeLogAbsDeterminant(ImW, d);
 
         // Score = (n*d/2) * log(sigma^2) - n * log|det(I-W)|
-        double score = (n * d / 2.0) * Math.Log(sigmaSquared) - n * logDetImW;
+        // Same overflow: n * d is evaluated in int before /2.0 promotes the result.
+        double score = ((double)n * d / 2.0) * Math.Log(sigmaSquared) - n * logDetImW;
 
         // Gradient
         var grad = new Matrix<T>(d, d);
